+++ CHANGELOG.md
@@ -1319,3 +1319,5 @@ Alternately, the _Daily Diary of my Programming Dreams_.
 * **C**
     1.  Added [fiddles/c/fiddle-0001-DataTypes](fiddles/c/fiddle-0001-DataTypes) ~ [Issue 90](https://github.com/bradyhouse/house/issues/90)
     2.  Added [fiddles/c/fiddle-0002-ArithmeticOperators](fiddles/c/fiddle-0002-ArithmeticOperators) ~ [Issue 91](https://github.com/bradyhouse/house/issues/91)
+    3.  Added [fiddles/c/fiddle-0003-ModulusOperator](fiddles/c/fiddle-0003-ModulusOperator) ~ [Issue 92](https://github.com/bradyhouse/house/issues/92)
+